Moved medical center out of admin into models as it should be
