+++ submission/admin.py
@@ -4,7 +4,7 @@ import uuid
 
 from django.db import models
 from submission.models import Submission, Admission, PersonalData, Phone, OverallWellbeing, CommonSymptoms, \
-    GradedSymptoms, RelatedConditions
+    GradedSymptoms, RelatedConditions, MedicalCenter
 
 
 class AdmissionInline(admin.TabularInline):
@@ -48,19 +48,5 @@ class SubmissionAdmin(admin.ModelAdmin):
         RelatedConditionsInline
     ]
 
-class MedicalCenter(models.Model):
-    """
-    Available medical centers, from which a patient can choose.
-
-    This is as limited set of attributes, to reflect what is needed for a simple, 
-    GPS located submission.
-
-    Addtional address information could be added in the future if necessary.
-    """
-
-    id = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
-
-    name = models.CharField(max_length=50, blank=False)
-    latitude = models.FloatField(null=False)
-    longitude = models.FloatField(null=False)
-
+class MedicalCenterAdmin(models.Model):
+    model = MedicalCenter
