+++ submission/models.py
@@ -3,7 +3,6 @@ import uuid
 
 from django.core.validators import MaxValueValidator, MinValueValidator
 from django.db import models
-from submission.admin import MedicalCenter
 
 class Patient(models.Model):
     """
@@ -163,6 +162,23 @@ class RelatedConditions(models.Model):
     take_immunosuppressants = models.BooleanField(null=False, default=False)
     pregnant = models.BooleanField(null=False, default=False)
     smoke = models.BooleanField(null=False, default=False)
+
+
+class MedicalCenter(models.Model):
+    """
+    Available medical centers, from which a patient can choose.
+
+    This is as limited set of attributes, to reflect what is needed for a simple, 
+    GPS located submission.
+
+    Addtional address information could be added in the future if necessary.
+    """
+
+    id = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
+
+    name = models.CharField(max_length=50, blank=False)
+    latitude = models.FloatField(null=False)
+    longitude = models.FloatField(null=False)
 
 
 class ChosenMedicalCenter(models.Model):
